fix: separating the `websockets` code to its own domain
[ci skip]
